% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hc_eee_pc}
\alias{hc_eee_pc}
\title{Health Care Expenditures per Capita by State of Residence}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1224 rows and 5 columns.}
\source{
\url{https://www.kff.org/other/state-indicator/health-spending-per-capita}
}
\usage{
hc_eee_pc
}
\description{
Health Care Expenditures per Capita by State of Residence
}
\keyword{datasets}
