% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_em_visits}
\alias{hosp_em_visits}
\title{Hospital Emergency Room Visits per 1,000 Population by Ownership Type}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 867 rows and 8 columns.}
\source{
\url{https://www.kff.org/other/state-indicator/emergency-room-visits-by-ownership/}
}
\usage{
hosp_em_visits
}
\description{
Hospital Emergency Room Visits per 1,000 Population by Ownership Type
}
\keyword{datasets}
