% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_ip_days}
\alias{hosp_ip_days}
\title{Hospital Inpatient Days per 1,000 Population by Ownership Type}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 867 rows and 8 columns.}
\source{
\url{https://www.kff.org/other/state-indicator/inpatient-days-by-ownership/}
}
\usage{
hosp_ip_days
}
\description{
Hospital Inpatient Days per 1,000 Population by Ownership Type
}
\keyword{datasets}
