% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicaid_chip_enrollment}
\alias{medicaid_chip_enrollment}
\title{Total Monthly Medicaid and CHIP Enrollment}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3060 rows and 8 columns.}
\source{
\url{https://www.kff.org/health-reform/state-indicator/total-monthly-medicaid-and-chip-enrollment}
}
\usage{
medicaid_chip_enrollment
}
\description{
Total Monthly Medicaid and CHIP Enrollment
}
\keyword{datasets}
