% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_diabetes}
\alias{pct_diabetes}
\title{Percent of Adults Who Have Ever Been Told by a Doctor that They Have Diabetes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 204 rows and 8 columns.}
\source{
\url{https://www.kff.org/other/state-indicator/adults-with-diabetes/}
}
\usage{
pct_diabetes
}
\description{
Percent of Adults Who Have Ever Been Told by a Doctor that They Have Diabetes
}
\keyword{datasets}
