% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pct_overweight_obesity}
\alias{pct_overweight_obesity}
\title{Percent of Adults Who are Overweight or Obese}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 204 rows and 5 columns.}
\source{
\url{https://www.kff.org/other/state-indicator/adult-overweightobesity-rate}
}
\usage{
pct_overweight_obesity
}
\description{
Percent of Adults Who are Overweight or Obese
}
\keyword{datasets}
